requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
cryptography==2.9
decorator==4.4.2
future==0.18.2
pycparser==2.20
python-telegram-bot==12.6.1
pytz==2020.1
six==1.14.0
tornado==6.0.4
